+++ view_create.html
@@ -11,7 +11,7 @@
 <h1>
     新規作成
 </h1>
-<form method="post" action="model.php">
+<form method="post" action="controller.php">
     <div style="margin: 10px">
         <label for="title">タイトル：</label>
         <input id="title" type="text" name="title">
@@ -20,11 +20,11 @@
         <label for="content">内容：</label>
         <textarea id="content" name="content" rows="8" cols="40"></textarea>
     </div>
-    <input name="request" type="hidden" value="create">
+    <input name="request" type="hidden" value="todoCreate">
     <input type="submit" name="post" value="投稿する">
 </form>
-<form action="index.php">
-    <button type="submit" name="back">戻る</button>
+<form method="post" action="controller.php">
+    <button type="submit" name="request" value="todoAscend">戻る</button>
 </form>
 </body>
 </html>
