+++ view_edit.php
@@ -1,3 +1,12 @@
+<?php
+session_start();
+session_regenerate_id(true);
+include_once('model_Sanitization.php');
+$sanitizer = new Sanitization();
+$default = $sanitizer->sanitize($_SESSION['default'][0]);
+session_unset();
+?>
+
 <!doctype html>
 <html lang="ja">
 <head>
@@ -9,28 +18,25 @@
 </head>
 <body>
 
-<?php
-//DBで受け継いだデータを$recに格納、もしくは
-?>
+
 <h1>
     Edit Todo Page
 </h1>
-<form method="post" action="model.php">
+<form method="post" action="controller.php">
     <div style="margin: 10px">
         <label for="title">タイトル：</label>
-        <input id="title" type="text" name="title" value="<?php print $default['title']?>">
+        <input id="title" type="text" name="title" value="<?php echo $default['title']?>">
     </div>
     <div style="margin: 10px">
         <label for="content">内容：</label>
-        <textarea id="content" name="content" rows="8" cols="40"><?php print $default['content']?></textarea>
+        <textarea id="content" name="content" rows="8" cols="40"><?php echo $default['content']?></textarea>
     </div>
-    <input name="id" type="hidden" value="<?php print $default['id'];?>">
+    <input name="id" type="hidden" value="<?php echo $default['id'];?>">
     <input name="request" type="hidden" value="todoUpdate">
     <input type="submit" name="post" value="完了">
 </form>
 <form method="post" action="controller.php">
-    <input name="request" type="hidden" value="todoAscend">
-    <button type="submit" name="back">戻る</button>
+    <button type="submit" name="request" value="todoAscend">戻る</button>
 </form>
 </body>
 </html>
